refactor: Move all stores to source/pinia
