+++ source/pinia/index.ts
@@ -14,3 +14,5 @@
 
 export { useWorkspacesStore } from './workspaces'
 export { useOpenDirectoryStore } from './open-directory'
+export { useConfigStore } from './config'
+export { useStatisticsStore } from './statistics-store'
